# Conflicts:
#	dcat_schema_transpiler/dcat_schema_transpiler/ckan_schema/mobility_dcat_ap_converter/class_converter.py
#	dcat_schema_transpiler/dcat_schema_transpiler/ckan_schema/mobility_dcat_ap_converter/classes/dataset.py
#	dcat_schema_transpiler/dcat_schema_transpiler/mobility_dcat_ap/dataset.py
#	dcat_schema_transpiler/dcat_schema_transpiler/mobility_dcat_ap_to_schema.py
#	ext/ckanext-digitraffic_theme/ckanext/digitraffic_theme/model/dataset.py
#	ext/ckanext-digitraffic_theme/ckanext/digitraffic_theme/model/mobility_data.py
#	ext/ckanext-digitraffic_theme/ckanext/digitraffic_theme/profiles/mobility_profile.py
#	ext/ckanext-digitraffic_theme/ckanext/digitraffic_theme/schemas/mobility_dcat.yaml

+++ local-env/start-local-ckan.sh
@@ -39,7 +39,9 @@ run_python_install() {
   fi
   source venv/bin/activate
   pip install -e .
-  pip install -r requirements.txt
+  if [[ -f requirements.txt ]]; then
+    pip install -r requirements.txt
+  fi
   deactivate
   popd
 }
